+++ src/fedora_system_monitor/__init__.py
@@ -1,3 +1,3 @@
 """Fedora System Monitor."""
 
-__version__ = "1.0.1"
+__version__ = "1.1.0"